python: split flux.job module into multiple files
Problem: The Python 'flux.job' module is large (>1200 lines)
and thus getting difficult to read and/or modify. This module will
only get larger since there will certainly be more classes and
methods related to Flux jobs added to the Python API.

Split `src/bindings/python/flux/job.py` into multiple files, while
attempting to keep related code together without making any one component
too large.

For backwards compatibility, recreate the previous `flux.job` namespace
in the __init__.py in the new job directory.
