+++ src/bindings/python/flux/job/__init__.py
@@ -1,0 +1,26 @@
+###############################################################
+# Copyright 2020 Lawrence Livermore National Security, LLC
+# (c.f. AUTHORS, NOTICE.LLNS, COPYING)
+#
+# This file is part of the Flux resource manager framework.
+# For details, see https://github.com/flux-framework.
+#
+# SPDX-License-Identifier: LGPL-3.0
+###############################################################
+
+from flux.job.Jobspec import Jobspec, JobspecV1, validate_jobspec
+from flux.job.JobID import id_parse, id_encode, JobID
+from flux.job.kvs import job_kvs, job_kvs_guest
+from flux.job.kill import kill_async, kill, cancel_async, cancel
+from flux.job.submit import submit_async, submit, submit_get_id
+from flux.job.list import job_list, job_list_inactive, job_list_id
+from flux.job.wait import wait_async, wait, wait_get_status
+from flux.job.event import (
+    event_watch_async,
+    event_watch,
+    event_wait,
+    JobEventWatchFuture,
+    EventLogEvent,
+    JobException,
+)
+from flux.core.inner import ffi
