+++ src/bindings/python/flux/job/_wrapper.py
@@ -1,0 +1,20 @@
+###############################################################
+# Copyright 2020 Lawrence Livermore National Security, LLC
+# (c.f. AUTHORS, NOTICE.LLNS, COPYING)
+#
+# This file is part of the Flux resource manager framework.
+# For details, see https://github.com/flux-framework.
+#
+# SPDX-License-Identifier: LGPL-3.0
+###############################################################
+
+from flux.wrapper import Wrapper
+from _flux._core import ffi, lib
+
+
+class JobWrapper(Wrapper):
+    def __init__(self):
+        super(JobWrapper, self).__init__(ffi, lib, prefixes=["flux_job_"])
+
+
+_RAW = JobWrapper()
